epa::ppx_luminosity_b*: improve calculation accuracy
Remove the leading 1 from ppx_luminosity_internal. It is integrated
separately when the spectrum with non-electromagnetic interactions is
provided. Before this patch, it was subtracted from the result of
ppx_luminosity_internal. It resulted in a loss of accuracy and eventually
a waste of time when the 1 was the dominating contribution.